refactor: [metrics] adjust domains service
